@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.1.0+31
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -28,26 +28,29 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
-  dio: ^3.0.10
-  provider: ^4.3.2+2
-  cached_network_image: ^2.3.2+1
-  url_launcher: ^5.7.2
-  shared_preferences: ^0.5.12
-  carousel_slider: ^2.3.1
-  google_maps_place_picker: ^0.10.0
-  # flutter_email_sender: ^3.0.1
+  cupertino_icons: ^1.0.4
+  dio: ^4.0.6
+  flutter_google_places_hoc081098: ^1.0.0-nullsafety.5
+  google_api_headers: ^1.1.1
+  geocoding: ^2.0.0
+
+  google_maps_webservice: ^0.0.20-nullsafety.5
+  provider: ^6.0.0
+  cached_network_image: ^3.2.0
+  url_launcher: ^6.0.20
+  shared_preferences: ^2.0.13
+  carousel_slider: ^4.0.0
   youtube_player_flutter: ^8.0.0
-  geolocator: ^6.0.0+4
-  google_maps_flutter: ^0.5.25+3
-  map_launcher: ^1.1.2
-  video_player:
-  ffi: ^1.0.0
-  getwidget: ^2.0.4
-  flutter_tawk: ^0.0.5
-  flutter_html:
+  geolocator: ^7.7.1
+  google_maps_flutter: ^2.1.3
+  map_launcher: ^2.2.1+2
+  video_player: ^2.3.0
+  ffi: ^1.1.2
+  getwidget: ^2.0.5
+  flutter_tawk: ^0.1.0
+  flutter_html: ^3.0.0-alpha.2
   lazy_load_scrollview: ^1.3.0
-  group_radio_button:
+  group_radio_button: ^1.2.0
 
 
 
