@@ -5,9 +5,9 @@ class ProductSize {
     this.titleen,
   });
 
-  String sizeId;
-  String titlear;
-  String titleen;
+  String? sizeId;
+  String? titlear;
+  String? titleen;
 
   factory ProductSize.fromJson(Map<String, dynamic> json) => ProductSize(
         sizeId: json["size_id"] == null ? null : json["size_id"],
