@@ -4,7 +4,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 class AppLanguage extends ChangeNotifier {
   Locale _appLocale = Locale('ar');
 
-  Locale get appLocal => _appLocale ?? Locale("ar");
+  Locale get appLocal => _appLocale ;
 
   fetchLocale() async {
     var prefs = await SharedPreferences.getInstance();
@@ -12,7 +12,7 @@ class AppLanguage extends ChangeNotifier {
       _appLocale = Locale('ar');
       return Null;
     }
-    _appLocale = Locale(prefs.getString('language_code'));
+    _appLocale = Locale(prefs.getString('language_code')!);
     return Null;
   }
 
