@@ -5,7 +5,7 @@ import 'package:fbTrade/ui/home_screen.dart';
 import 'package:flutter/material.dart';
 
 class NewViewOne extends StatefulWidget {
-  List<HomeCategory> list = [];
+  List<HomeCategory>? list = [];
 
   NewViewOne({this.list});
   @override
@@ -23,12 +23,12 @@ class _NewViewOneState extends State<NewViewOne> {
         primary: false,
         shrinkWrap: true,
         physics: NeverScrollableScrollPhysics(),
-        itemCount: widget.list.length,
+        itemCount: widget.list!.length,
         itemBuilder: (context, index) {
           return InkWell(
             onTap: () {
               Navigator.of(context).push(MaterialPageRoute(
-                builder: (context) => CategoryDetails(widget.list[index],
+                builder: (context) => CategoryDetails(widget.list![index],
                     Localizations.localeOf(context).languageCode),
               ));
             },
@@ -47,8 +47,8 @@ class _NewViewOneState extends State<NewViewOne> {
                 ]),
                 child: CachedNetworkImage(
                   imageUrl: Localizations.localeOf(context).languageCode == "en"
-                      ? "${widget.list[index].picpathEn}"
-                      : "${widget.list[index].picpath}",
+                      ? "${widget.list![index].picpathEn}"
+                      : "${widget.list![index].picpath}",
                   fit: BoxFit.cover,
                 ),
               ),
